backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-firestore
google-auth